worker/requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
tenacity>=8.2.0
